rename hjelpetekstVurdereMålgruppe
